Split oversized local bridge test files
