+++ .pm/tasks/task_0820814da7d849159673dc3fece649e2.execution.md
@@ -686,3 +686,12 @@ Example:
 - Expected Result: PR exists on GitHub, references task_0820814da7d849159673dc3fece649e2, and is ready for normal PR CI/comment/mergeability watch.
 - Actual Result: PASS; PR URL: https://github.com/eng-cc/oasis7/pull/421 .
 - Blocker / Next Action: No blocker; push this PR evidence commit and watch PR #421.
+
+## 2026-06-12 18:15:17 CST / tpm
+- 完成内容: Fixed PR #421 required-gate failure caused by Rust oversized-file guard.
+- 遗留事项: Push the CI fix and re-watch PR #421 checks/comments/mergeability.
+- Action: Split provider bridge model-decision/prompt/guardrail helpers into crates/oasis7/src/bin/oasis7_provider_local_bridge/agent_decision.rs and split launcher viewer static dir tests into crates/oasis7/src/bin/oasis7_game_launcher/viewer_static_dir_tests.rs. This keeps original files below the 1200-line gate without changing behavior.
+- Validation Command: ./scripts/check-rust-file-size.sh && env -u RUSTC_WRAPPER cargo fmt --all --check; env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_provider_local_bridge -- --nocapture; ./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_game_launcher provider_backed_viewer_live_env_preserves_local_mock_backend -- --nocapture && ./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_game_launcher build_viewer_live_command_wires_auto_play_flag -- --nocapture && ./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_game_launcher viewer_static_dir -- --nocapture; git diff --check
+- Expected Result: Rust file-size gate reports zero oversized files, moved provider/launcher tests still pass, formatting and whitespace gates pass.
+- Actual Result: PASS; check-rust-file-size OK with oversized code files=0 and test files=0; provider local bridge bin tests passed 23/23; launcher focused tests passed including 4 viewer_static_dir tests; cargo fmt --check and git diff --check passed.
+- Blocker / Next Action: No local blocker; commit and push the CI fix, then watch PR #421 required checks again.
